+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 - refactor: remove RxJS dependency
 - feat: cursor back/forward
 - feat: history improvements
+- feat: max history configuration
 
 ## [1.0.0] - 2021-02-06